quicksave behavior tweak

+++ Blender/2.69/scripts/addons/cc_utils_misc.py
@@ -14,11 +14,14 @@ bl_info = {
 _save_iter = 0
 _save_path = None
 _save_lock = False
+_save_main = False
 
 def save_blendfile_copy(path):
     global _save_lock
+    global _save_main
 
     _save_lock = True
+    _save_main = False
     bpy.ops.wm.save_as_mainfile(
         filepath=path,
         check_existing=False,
@@ -58,8 +61,14 @@ def save_scratch_blendfile(dirpath=None, dirty_only=False):
     return path
 
 def count_scratch_iterations(path):
-    dirname = os.path.dirname(path)
-    return len([name for name in os.listdir(dirname) if name.endswith('.blend')])
+    if path.endswith('.blend'):
+        dirname = os.path.dirname(path)
+    else:
+        dirname = path
+    if os.path.isdir(dirname):
+        return len([name for name in os.listdir(dirname) if name.endswith('.blend')])
+    else:
+        return 0
 
 def update_backups_prop(self, context):
     for scene in bpy.data.scenes:
@@ -69,22 +78,28 @@ def update_backups_prop(self, context):
 PROP_BACKUPS = bpy.props.BoolProperty(name='Backups', update=update_backups_prop)
 
 @bpy.app.handlers.persistent
-def load_pre(scene):
+def load_post(scene):
     global _save_path
     global _save_iter
     _save_path = None
     _save_iter = 0
+    path = get_scratch_path()
+    if os.path.isdir(path):
+        _save_path = path
+        _save_iter = count_scratch_iterations(path)
 
 @bpy.app.handlers.persistent
 def save_post(scene):
     global _save_lock
     global _save_path
     global _save_iter
+    global _save_main
     if not _save_lock and bpy.context.scene.backups:
-        path = save_scratch_blendfile(dirty_only=True)
-        if path:
-            _save_path = path
-            _save_iter = count_scratch_iterations(path)
+        filename = save_scratch_blendfile(dirty_only=True)
+        if filename:
+            _save_path = os.path.dirname(filename)
+            _save_iter = count_scratch_iterations(_save_path)
+        _save_main = True
 
 class Quicksave(bpy.types.Operator):
     bl_idname = 'cc.quicksave'
@@ -94,11 +109,12 @@ class Quicksave(bpy.types.Operator):
     def execute(self, context):
         global _save_path
         global _save_iter
-        path = save_scratch_blendfile(dirty_only=True)
-        if path:
-            _save_path = path
-            _save_iter = count_scratch_iterations(path)
-            self.report({'INFO'}, 'Quicksave: %s' % os.path.basename(path))
+        global _save_main
+        filename = save_scratch_blendfile(dirty_only=True and not _save_main)
+        if filename:
+            _save_path = os.path.dirname(filename)
+            _save_iter = count_scratch_iterations(_save_path)
+            self.report({'INFO'}, 'Quicksave: %s' % os.path.basename(filename))
         return {'FINISHED'}
 
 class DeleteScratch(bpy.types.Operator):
@@ -110,13 +126,14 @@ class DeleteScratch(bpy.types.Operator):
         global _save_path
         global _save_iter
         path = get_scratch_path()
-        names = [name for name in os.listdir(path) if name.endswith('.blend')]
-        for name in names:
-            filename = os.path.join(path, name)
-            os.remove(filename)
-        _save_path = ''
-        _save_iter = 0
-        self.report({'INFO'}, 'Scratch files deleted: %s' % len(names))
+        if os.path.isdir(path):
+            names = [name for name in os.listdir(path) if name.endswith('.blend')]
+            for name in names:
+                filename = os.path.join(path, name)
+                os.remove(filename)
+            _save_path = ''
+            _save_iter = 0
+            self.report({'INFO'}, 'Scratch files deleted: %s' % len(names))
         return {'FINISHED'}
 
 def quicksave_info(panel, context):
@@ -124,7 +141,7 @@ def quicksave_info(panel, context):
     global _save_iter
     layout = panel.layout
     if _save_path:
-        layout.label('(%s) %s' % (_save_iter, os.path.dirname(_save_path)))
+        layout.label('(%s) %s' % (_save_iter, _save_path))
 
 def cardboard_menu_ext(self, context):
     self.layout.prop(context.scene, 'backups')
@@ -133,7 +150,7 @@ def register():
     cc.utils.register(__REGISTER__)
     cc.ui.CardboardMenu.add_section(cardboard_menu_ext, 9000)
 
-    bpy.app.handlers.load_pre.append(load_pre)
+    bpy.app.handlers.load_post.append(load_post)
     bpy.app.handlers.save_post.append(save_post)
     bpy.types.INFO_HT_header.append(quicksave_info)
 
@@ -141,7 +158,7 @@ def unregister():
     cc.utils.unregister(__REGISTER__)
     cc.ui.CardboardMenu.remove_section(cardboard_menu_ext)
 
-    bpy.app.handlers.load_pre.remove(load_pre)
+    bpy.app.handlers.load_post.remove(load_post)
     bpy.app.handlers.save_post.remove(save_post)
     bpy.types.INFO_HT_header.remove(quicksave_info)
